Crawler/reddit_crawler/crawler/requirements.txt
requests
psycopg2-binary
tenacity
python-dotenv
pyfaktory